requirements.txt
Click==8.0.1
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dotenv==0.15.0
Werkzeug==1.0.1
